touchscreen calibration now done with EEPROM! Automatically loads cal values from eeprom instead of using defaults if the eeprom is empty or faulted, loads the defaults this means that we can now add a button on the steering wheel somewhere to recalibrate the touchscreen, and those values will actually be stored!
Also made MasterConfig.h and started moving some #defines there. This
file will basically contain all the settings that you might want to
change

Added preprocessor-controlled debug blocks. If you want to debug, set
the appropriate values in MasterConfig.h and the debug code will be
compiled. If you don't want to debug, the code won't even be compiled!

addresses #14, started working on doing the same as above but for launch
profiles. Stopepd because Anastasia wanted to work on it.